requirements.txt
python-binance==1.0.29
ta==0.10.2
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.1
